updates from in-person code review
