+++ src/main/scala/org/broadinstitute/dsde/firecloud/dataaccess/SearchDAO.scala
@@ -3,6 +3,6 @@ package org.broadinstitute.dsde.firecloud.dataaccess
 /**
   * Created by davidan on 9/28/16.
   */
-abstract class SearchDAO {
+trait SearchDAO {
 
 }
