+++ src/main/scala/org/broadinstitute/dsde/firecloud/utils/RoleSupport.scala
@@ -16,18 +16,6 @@ trait RoleSupport {
   protected val userInfo: UserInfo
   implicit protected val executionContext: ExecutionContext
 
-  /*
-  def tryIsFCAdmin(userId: String): Future[Boolean] = {
-    gcsDAO.isAdmin(userId) recoverWith { case t => throw new RawlsException("Unable to query for admin status.", t) }
-  }
-
-  def asFCAdmin(op: => Future[PerRequestMessage]): Future[PerRequestMessage] = {
-    tryIsFCAdmin(userInfo.userEmail) flatMap { isAdmin =>
-      if (isAdmin) op else Future.failed(new RawlsExceptionWithErrorReport(errorReport = ErrorReport(StatusCodes.Forbidden, "You must be an admin.")))
-    }
-  }
-  */
-  
   def tryIsCurator(userInfo: UserInfo): Future[Boolean] = {
     rawlsDAO.isLibraryCurator(userInfo) recoverWith { case t => throw new FireCloudException("Unable to query for library curator status.", t) }
   }
