+++ src/test/scala/org/broadinstitute/dsde/firecloud/service/LibraryServiceSpec.scala
@@ -1,11 +1,14 @@
 package org.broadinstitute.dsde.firecloud.service
 
+import org.broadinstitute.dsde.firecloud.dataaccess.RawlsDAO
 import org.broadinstitute.dsde.firecloud.model.AttributeUpdateOperations._
-import org.broadinstitute.dsde.firecloud.model.AttributeString
+import org.broadinstitute.dsde.firecloud.model.{AttributeString, UserInfo}
 import org.scalatest.FreeSpec
 import spray.json._
 
-class LibraryServiceSpec extends FreeSpec {
+import scala.concurrent.ExecutionContext
+
+class LibraryServiceSpec extends FreeSpec with LibraryServiceSupport {
 
   val existingAttrs1 = Map("library:keyone"->"valone", "library:keytwo"->"valtwo", "library:keythree"->"valthree", "library:keyfour"->"valfour")
 
@@ -20,7 +23,7 @@ class LibraryServiceSpec extends FreeSpec {
           RemoveAttribute("library:keyfour")
         )
         assertResult(expected) {
-          LibraryService.generateAttributeOperations(existingAttrs1, newAttrs.asJsObject)
+          generateAttributeOperations(existingAttrs1, newAttrs.asJsObject)
         }
       }
     }
@@ -34,7 +37,7 @@ class LibraryServiceSpec extends FreeSpec {
           AddUpdateAttribute("library:keytwo",AttributeString("valtwoNew"))
         )
         assertResult(expected) {
-          LibraryService.generateAttributeOperations(existingAttrs1, newAttrs.asJsObject)
+          generateAttributeOperations(existingAttrs1, newAttrs.asJsObject)
         }
       }
     }
